backend/requirements/base.txt
Django==2.0

mysqlclient==1.3.9
django-cors-headers==2.0.0
requests==2.12.4

# Time zones support
pytz==2016.10

Pillow
django-model-utils
djangorestframework==3.7.7
djangorestframework_simplejwt